Provide dedicated implementations for \ArrayAccess to minimize warnings because if mismatched function declarations
